requirements.txt
absl-py==2.3.1
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==4.4.2
et_xmlfile==2.0.0
ffmpeg==1.4
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.53.1
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h5py==3.14.0
httplib2==0.22.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
kiwisolver==1.4.7
kombu==5.5.4
libclang==18.1.1
llvmlite==0.44.0
lxml==5.4.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
mysql-connector-python==9.3.0
namex==0.1.0
narwhals==1.48.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.1.3
openai-whisper==20240930
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.17.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
proglog==0.1.12
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rich==14.1.0
rpds-py==0.26.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
setuptools==80.7.1
six==1.16.0
smmap==5.0.2
soupsieve==2.7
SpeechRecognition==3.14.3
statsmodels==0.14.2
streamlit==1.47.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
threadpoolctl==3.5.0
tiktoken==0.9.0
toml==0.10.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
whisper==1.1.10
wrapt==1.17.2
xgboost==3.0.5